- Add CountablePageIterator - Rename classes
