--- src/CountablePageIterator.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace ShoppingFeed\Paginator;
-
-use ShoppingFeed\Iterator\CountableTraversable;
-
-class CountablePageIterator implements CountableTraversable
-{
-
-}
